+++ imports/api/unlocations/unlocationsCollection.js
@@ -1,4 +1,3 @@
 import { Mongo } from 'meteor/mongo';
 
 export const UNLocations = new Mongo.Collection('UNLocations');
-