+++ imports/ui/fields/UNLocationField.jsx
@@ -31,7 +31,7 @@ const UNLocationField = (props) => {
           const options = res.map(opt => ({
             value: opt._id,
             label: opt.name,
-            isPort: opt.isPort,
+            isSeaport: opt.isSeaport,
             isAirport: opt.isAirport,
           }));
           cb(null, { options });
@@ -43,7 +43,7 @@ const UNLocationField = (props) => {
   const locationRenderer = option => (
     <div>
       {
-        option.isPort ?
+        option.isSeaport ?
         (
           <span><span className="fa fa-fw fa-ship" />&nbsp;</span>
         ) : ''}
@@ -53,7 +53,7 @@ const UNLocationField = (props) => {
           <span><span className="fa fa-fw fa-plane" />&nbsp;</span>
         ) : ''}
       {
-        !option.isPort && !option.isAirport ?
+        !option.isSeaport && !option.isAirport ?
         (
           <span><span className="fa fa-fw" />&nbsp;</span>
         ) : ''}