+++ private/importAirports.js
@@ -1,0 +1,11 @@
+var db = connect('localhost:3001/meteor');
+
+var airports = db.UNLocations.find({ isAirport: true });
+
+while (airports.hasNext()) {
+  var airport = airports.next();
+  var id = airport._id;
+  airport._id = ObjectId().str;
+  db.UNLocations.insert(airport);
+  db.UNLocations.update({ _id: id }, { $set: { isAirport: false } });
+}